Event persistence
This commit adds a simple SQLite-based persistence of events. Now, if
the bridge is stopped and later restarted, orders that were previously
in progress will be resurrected and put back into their correct place
in the workflow.

We store all the orders in one event log table, and then select only
the latest ones when we want to retrieve the current set for a given
state (basically, we’re doing simple versioning here).