Make sure we keep up to date with current block state
If a long period elapses between receiving events, we currently don't
update the block we are querying against. As there is a maximum limit
to how far in the past we can query for blocks, this eventually means
that we hit an error condition because our block counter isn't up to
date and we are asking for too far behind.

So now, ask for the current block when we make a call, and use this as
the default for the next run. This means our block tracker stays up to
date with newly minted blocks.