Upgrade bridge to include configurable result type
We now support specifying in the job call what type of result the
caller would like. This is to support jobs that write data directly to
STDOUT or STDERR to then be usable in the contract directly.

The reason we don't just return this info for all jobs is that some
jobs will produce a lot of log output, and sending this all back to
the contract will cost the user a ton of gas. So sometimes you just
want the result CID and sometimes you want STDOUT.